+++ packages/config/src/types/integrations.ts
@@ -1,0 +1,8 @@
+export type Integration = {
+  slug: string
+  version?: string
+  has_build?: boolean
+  dev?: {
+    path: string
+  }
+}